Update calc_edit_asl.adoc
I used a combination of edits and // comments to suggest changes. Some suggestions were due to me not knowing how experienced you expect your audience to be.
This looks like a good step-by-step presentation.
If I was to get pedantic, there are a lot of run-on sentences. If you want me to clean these up I could, but they aren't egregious.